+++ shared/types/view-type.ts
