+++ shared/ui/info-toast/index.tsx
@@ -2,8 +2,8 @@
 
 import CheckIcon from "@/shared/assets/icon/check.svg";
 import LockIcon from "@/shared/assets/icon/lock.svg";
+import { useToast } from "@/shared/hooks/use-toast";
 import { cn } from "@/shared/utils/cn";
-import { useEffect, useState } from "react";
 import * as styles from "./info-toast.css";
 
 type InfoToastProps = {
@@ -12,14 +12,7 @@ type InfoToastProps = {
 };
 
 const InfoToast = ({ status, infoText }: InfoToastProps) => {
-  const [exiting, setExiting] = useState(false);
-
-  useEffect(() => {
-    const exitTimer = setTimeout(() => setExiting(true), 3000);
-    return () => {
-      clearTimeout(exitTimer);
-    };
-  }, []);
+  const { exiting } = useToast();
 
   return (
     <div className={cn(styles.container, exiting ? styles.exit : styles.enter)}>
