+++ shared/ui/view-type-tabs/index.tsx
@@ -4,7 +4,7 @@ import GridIcon from "@/shared/assets/icon/grid.svg";
 import LayerIcon from "@/shared/assets/icon/layers.svg";
 import { cn } from "@/shared/utils/cn";
 
-import { ViewType } from "@/shared/ui/view-type-tabs/view-type";
+import { ViewType } from "@/shared/types/view-type";
 import * as styles from "./view-type-tabs.css";
 
 interface ViewTypeTabsProps {
